Refactoring: introduce src/bin/pgcopydb/progress.c.
