+++ src/bin/pgcopydb/copydb.h
@@ -6,6 +6,8 @@
 #ifndef COPYDB_H
 #define COPYDB_H
 
+#include "parson.h"
+
 #include "filtering.h"
 #include "lock_utils.h"
 #include "pgcmd.h"
@@ -375,10 +377,6 @@ bool copydb_collect_finished_subprocesses(bool *allDone);
 
 bool copydb_copy_roles(CopyDataSpec *copySpecs);
 
-bool copydb_prepare_schema_json_file(CopyDataSpec *copySpecs);
-bool copydb_parse_schema_json_file(CopyDataSpec *copySpecs);
-bool copydb_update_progress(CopyDataSpec *copySpecs, CopyProgress *progress);
-
 /* indexes.c */
 bool copydb_init_indexes_paths(CopyFilePaths *cfPaths,
 							   SourceIndexArray *indexArray,
@@ -485,5 +483,14 @@ bool copydb_copy_blobs(CopyDataSpec *specs);
 /* summary.c */
 bool prepare_summary_table(Summary *summary, CopyDataSpec *specs);
 bool print_summary(Summary *summary, CopyDataSpec *specs);
+
+/* progress.c */
+bool copydb_prepare_schema_json_file(CopyDataSpec *copySpecs);
+bool copydb_parse_schema_json_file(CopyDataSpec *copySpecs);
+bool copydb_update_progress(CopyDataSpec *copySpecs, CopyProgress *progress);
+
+bool copydb_progress_as_json(CopyDataSpec *copySpecs,
+							 CopyProgress *progress,
+							 JSON_Value *js);
 
 #endif  /* COPYDB_H */
